Move JavaFunc into protected sagemaker package
