+++ java/mlflow-java/src/main/java/com/databricks/mlflow/sagemaker/JavaFuncFlavor.java
@@ -1,4 +1,4 @@
-package com.databricks.mlflow.javafunc;
+package com.databricks.mlflow.sagemaker;
 
 import com.databricks.mlflow.Flavor;
